Refactors solution to practice inheritance.
