+++ high_fivable.rb
@@ -1,0 +1,5 @@
+module HighFivable
+	def offer_high_five
+    "High five!"
+  end
+end
